+++ src/containers/User/Registration.scss
